Rename project to omxplayer
