+++ omxplayer/__init__.py
@@ -1,0 +1,1 @@
+from omxplayer.player import OMXPlayer
