--- pyomxplayerng/__init__.py
@@ -1,1 +1,0 @@
-from pyomxplayerng.omxplayer import OMXPlayer
